requirements.txt
chromadb==0.5.5
sentence-transformers==3.0.1
numpy>=1.26
scikit-learn>=1.3
python-dotenv>=1.0
PyPDF2>=3.0
python-docx>=1.1
Flask>=3.0


